requirements.txt
fastapi>=0.115.0
matplotlib>=3.8.0
numpy>=1.26.0
pdfplumber>=0.11.0
pydantic>=2.9.0
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
uvicorn>=0.30.0
